+++ lib/shared/src/configuration.ts
@@ -40,10 +40,15 @@ export interface Configuration {
     autocompleteCompleteSuggestWidgetSelection?: boolean
     autocompleteExperimentalSyntacticPostProcessing?: boolean
     autocompleteExperimentalDynamicMultilineCompletions?: boolean
+    autocompleteExperimentalHotStreak?: boolean
     autocompleteExperimentalGraphContext: 'lsp-light' | 'bfg' | 'bfg-mixed' | null
     isRunningInsideAgent?: boolean
     agentIDE?: 'VSCode' | 'JetBrains' | 'Neovim' | 'Emacs'
     autocompleteTimeouts: AutocompleteTimeouts
+
+    testingLocalEmbeddingsModel: string | undefined
+    testingLocalEmbeddingsEndpoint: string | undefined
+    testingLocalEmbeddingsIndexLibraryPath: string | undefined
 }
 
 export interface AutocompleteTimeouts {