+++ vscode/src/chat/chat-view/SimpleChatModel.ts
@@ -5,6 +5,7 @@ import { TranscriptJSON } from '@sourcegraph/cody-shared/src/chat/transcript'
 import { InteractionJSON } from '@sourcegraph/cody-shared/src/chat/transcript/interaction'
 import { errorToChatError, InteractionMessage } from '@sourcegraph/cody-shared/src/chat/transcript/messages'
 import { reformatBotMessageForChat } from '@sourcegraph/cody-shared/src/chat/viewHelpers'
+import { ContextFileSource } from '@sourcegraph/cody-shared/src/codebase-context/messages'
 import { Message } from '@sourcegraph/cody-shared/src/sourcegraph-api'
 
 import { contextItemsToContextFiles } from './chat-helpers'
@@ -29,7 +30,8 @@ export class SimpleChatModel {
     constructor(
         public modelID: string,
         private messagesWithContext: MessageWithContext[] = [],
-        public readonly sessionID: string = new Date(Date.now()).toUTCString()
+        public readonly sessionID: string = new Date(Date.now()).toUTCString(),
+        public chatTitle?: string
     ) {}
 
     public isEmpty(): boolean {
@@ -111,6 +113,10 @@ export class SimpleChatModel {
         return this.messagesWithContext
     }
 
+    public setChatTitle(title: string): void {
+        this.chatTitle = title
+    }
+
     /**
      * Serializes to the legacy transcript JSON format
      */
@@ -124,6 +130,7 @@ export class SimpleChatModel {
         return {
             id: this.sessionID,
             chatModel: this.modelID,
+            chatTitle: this.chatTitle,
             lastInteractionTimestamp: this.sessionID,
             interactions,
         }
@@ -160,6 +167,7 @@ export interface ContextItem {
     uri: vscode.Uri
     range?: vscode.Range
     text: string
+    source?: ContextFileSource
 }
 
 export function contextItemId(contextItem: ContextItem): string {