+++ vscode/src/chat/chat-view/SimpleChatPanelProvider.ts
@@ -197,7 +197,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         }
 
         const viewType = CodyChatPanelViewType
-        const panelTitle = getChatPanelTitle(lastQuestion)
+        const panelTitle = this.history.getChat(this.sessionID)?.chatTitle || getChatPanelTitle(lastQuestion)
         const viewColumn = activePanelViewColumn || vscode.ViewColumn.Beside
         const webviewPath = vscode.Uri.joinPath(this.extensionUri, 'dist', 'webviews')
         const panel = vscode.window.createWebviewPanel(
@@ -235,7 +235,7 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         }
 
         const webviewPath = vscode.Uri.joinPath(this.extensionUri, 'dist', 'webviews')
-        panel.iconPath = vscode.Uri.joinPath(this.extensionUri, 'resources', 'cody.png')
+        panel.iconPath = vscode.Uri.joinPath(this.extensionUri, 'resources', 'active-chat-icon.svg')
 
         // Reset the webview options to ensure localResourceRoots is up-to-date
         panel.webview.options = {
@@ -448,6 +448,12 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         // and can be removed once we retire the old ChatPanelProvider
         return Promise.resolve()
     }
+    public handleChatTitle(title: string): void {
+        this.chatModel.setChatTitle(title)
+        if (this.webviewPanel) {
+            this.webviewPanel.title = title
+        }
+    }
 
     public triggerNotice(notice: { key: string }): void {
         void this.postMessage({
@@ -501,7 +507,28 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
         await this.saveSession(text)
         // trigger the context progress indicator
         this.postViewTranscript({ speaker: 'assistant' })
-        await this.generateAssistantResponse(requestID, userContextFiles, addEnhancedContext)
+        await this.generateAssistantResponse(requestID, userContextFiles, addEnhancedContext, contextSummary => {
+            if (submitType !== 'user') {
+                return
+            }
+
+            const properties = {
+                requestID,
+                chatModel: this.chatModel.modelID,
+                promptText: text,
+                contextSummary,
+            }
+            telemetryService.log('CodyVSCodeExtension:recipe:chat-question:executed', properties, {
+                hasV2Event: true,
+            })
+            telemetryRecorder.recordEvent('cody.recipe.chat-question', 'executed', {
+                metadata: { ...contextSummary },
+            })
+        })
+        // Set the title of the webview panel to the current text
+        if (this.webviewPanel) {
+            this.webviewPanel.title = this.history.getChat(this.sessionID)?.chatTitle || getChatPanelTitle(text)
+        }
     }
 
     private async handleEdit(requestID: string, text: string): Promise<void> {
@@ -526,7 +553,8 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
     private async generateAssistantResponse(
         requestID: string,
         userContextFiles?: ContextFile[],
-        addEnhancedContext = true
+        addEnhancedContext = true,
+        sendTelemetry?: (contextSummary: {}) => void
     ): Promise<void> {
         try {
             const contextWindowBytes = 28000 // 7000 tokens * 4 bytes per token
@@ -553,6 +581,23 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
                 const warningMsg =
                     'Warning: ' + warnings.map(w => (w.trim().endsWith('.') ? w.trim() : w.trim() + '.')).join(' ')
                 this.postError(new Error(warningMsg))
+            }
+
+            if (sendTelemetry) {
+                // Create a summary of how many code snippets of each context source are being
+                // included in the prompt
+                const contextSummary: { [key: string]: number } = {}
+                for (const { source } of newContextUsed) {
+                    if (!source) {
+                        continue
+                    }
+                    if (contextSummary[source]) {
+                        contextSummary[source] += 1
+                    } else {
+                        contextSummary[source] = 1
+                    }
+                }
+                sendTelemetry(contextSummary)
             }
 
             this.postViewTranscript({ speaker: 'assistant' })
@@ -698,6 +743,11 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
             chatID: this.sessionID,
         })
 
+        const chatTitle = this.history.getChat(this.sessionID)?.chatTitle
+        if (chatTitle) {
+            this.handleChatTitle(chatTitle)
+            return
+        }
         // Update webview panel title to match the last message
         const text = this.chatModel.getLastHumanMessages()?.displayText
         if (this.webviewPanel && text) {
@@ -850,7 +900,13 @@ export class SimpleChatPanelProvider implements vscode.Disposable, IChatPanelPro
     private async postCodyCommands(): Promise<void> {
         const send = async (): Promise<void> => {
             await this.editor.controllers.command?.refresh()
-            const prompts = (await this.editor.controllers.command?.getAllCommands(true)) || []
+            const allCommands = await this.editor.controllers.command?.getAllCommands(true)
+            // HACK: filter out commands that make inline changes and /ask (synonymous with a generic question)
+            const prompts =
+                allCommands?.filter(([id]) => {
+                    return !['/edit', '/doc', '/test', '/ask'].includes(id)
+                }) || []
+
             void this.postMessage({
                 type: 'custom-prompts',
                 prompts,
@@ -1038,7 +1094,7 @@ class ContextProvider implements IContextProvider {
             return []
         }
         logDebug('SimpleChatPanelProvider', 'getEnhancedContext > searching local embeddings')
-        const contextItems = []
+        const contextItems: ContextItem[] = []
         const embeddingsResults = await this.localEmbeddings.getContext(text, NUM_CODE_RESULTS + NUM_TEXT_RESULTS)
         for (const result of embeddingsResults) {
             const uri = vscode.Uri.from({
@@ -1054,6 +1110,7 @@ class ContextProvider implements IContextProvider {
                 uri,
                 range,
                 text: result.content,
+                source: 'embeddings',
             })
         }
         return contextItems
@@ -1076,7 +1133,7 @@ class ContextProvider implements IContextProvider {
         }
 
         logDebug('SimpleChatPanelProvider', 'getEnhancedContext > searching remote embeddings')
-        const contextItems = []
+        const contextItems: ContextItem[] = []
         const embeddings = await this.embeddingsClient.search([repoId], text, NUM_CODE_RESULTS, NUM_TEXT_RESULTS)
         if (isError(embeddings)) {
             throw new Error(`Error retrieving embeddings: ${embeddings}`)
@@ -1097,6 +1154,7 @@ class ContextProvider implements IContextProvider {
                 uri,
                 range,
                 text: codeResult.content,
+                source: 'embeddings',
             })
         }
 
@@ -1114,6 +1172,7 @@ class ContextProvider implements IContextProvider {
                 uri,
                 range,
                 text: textResult.content,
+                source: 'embeddings',
             })
         }
 
@@ -1268,7 +1327,12 @@ async function newChatModelfromTranscriptJSON(json: TranscriptJSON, defaultModel
             ]
         }
     )
-    return new SimpleChatModel(json.chatModel || defaultModelID, (await Promise.all(messages)).flat(), json.id)
+    return new SimpleChatModel(
+        json.chatModel || defaultModelID,
+        (await Promise.all(messages)).flat(),
+        json.id,
+        json.chatTitle
+    )
 }
 
 export function deserializedContextFilesToContextItems(
@@ -1298,6 +1362,7 @@ export function deserializedContextFilesToContextItems(
             uri,
             range,
             text: text || '',
+            source: file.source,
         }
     })
 }