+++ vscode/src/commands/CommandRunner.ts
@@ -229,20 +229,16 @@ export function addSelectionToPrompt(prompt: string, code: string): string {
 function getDocCommandRange(
     editor: vscode.TextEditor,
     selection: vscode.Selection,
-    languageId?: string
+    languageId: string
 ): vscode.Selection {
     const startLine = languageId === 'python' ? selection.start.line + 1 : selection.start.line
-    const pos = new vscode.Position(startLine, 0)
+    const adjustedStartPosition = new vscode.Position(startLine, 0)
 
-    // move the current selection to the defined selection in the text editor document
-    if (editor) {
-        const visibleRange = editor.visibleRanges
-        // reveal the range of the selection minus 5 lines if visibleRange doesn't contain the selection
-        if (!visibleRange.some(range => range.contains(selection))) {
-            // reveal the range of the selection minus 5 lines
-            editor?.revealRange(selection, vscode.TextEditorRevealType.InCenter)
-        }
+    if (editor && !editor.visibleRanges.some(range => range.contains(adjustedStartPosition))) {
+        // reveal the range of the selection if visibleRange doesn't contain the selection
+        // we only use the start position as it is possible that the selection covers more than the entire visible area
+        editor.revealRange(selection, vscode.TextEditorRevealType.InCenter)
     }
 
-    return new vscode.Selection(pos, pos)
+    return new vscode.Selection(adjustedStartPosition, selection.end)
 }