+++ vscode/src/services/telemetry.ts
@@ -4,10 +4,10 @@ import { Configuration, ConfigurationWithAccessToken } from '@sourcegraph/cody-s
 import { TelemetryEventProperties, TelemetryService } from '@sourcegraph/cody-shared/src/telemetry'
 import { EventLogger, ExtensionDetails } from '@sourcegraph/cody-shared/src/telemetry/EventLogger'
 
-import { version as packageVersion } from '../../package.json'
 import { getConfiguration } from '../configuration'
 import { logDebug } from '../log'
 import { getOSArch } from '../os'
+import { version } from '../version'
 
 import { localStorage } from './LocalStorageProvider'
 
@@ -17,17 +17,12 @@ let globalAnonymousUserID: string
 
 const { platform, arch } = getOSArch()
 
-export const extensionVersion =
-    vscode.extensions.getExtension('sourcegraph.cody-ai')?.packageJSON?.version ?? packageVersion
 export const getExtensionDetails = (config: Pick<Configuration, 'agentIDE'>): ExtensionDetails => ({
     ide: config.agentIDE ?? 'VSCode',
     ideExtensionType: 'Cody',
     platform: platform ?? 'browser',
     arch,
-    // Prefer the runtime package json over the version that is inlined during build times. This
-    // way we will be able to include pre-release builds that are published with a different version
-    // identifier.
-    version: extensionVersion,
+    version,
 })
 
 /**