+++ vscode/webviews/Components/EnhancedContextSettings.tsx
@@ -114,6 +114,7 @@ const EmbeddingsConsentComponent: React.FunctionComponent<{ provider: LocalEmbed
 function contextProviderState(provider: ContextProvider): React.ReactNode {
     switch (provider.state) {
         case 'indeterminate':
+            return <></>
         case 'ready':
             if (provider.kind === 'embeddings' && provider.type === 'remote') {
                 return (
@@ -128,12 +129,28 @@ function contextProviderState(provider: ContextProvider): React.ReactNode {
         case 'unconsented':
             return <EmbeddingsConsentComponent provider={provider} />
         case 'no-match':
-            return (
-                <p className={styles.providerExplanatoryText}>
-                    {/* No repository matching {provider.remoteName} on <a href="about:blank#TODO">{provider.origin}</a> */}
-                    No repository matching {provider.remoteName} on {provider.origin}
-                </p>
-            )
+            if (provider.kind === 'embeddings') {
+                if (provider.type === 'remote') {
+                    return (
+                        <p className={styles.providerExplanatoryText}>
+                            No repository matching {provider.remoteName} on {provider.origin}
+                        </p>
+                    )
+                }
+                // Error messages for local embeddings missing.
+                switch (provider.errorReason) {
+                    case 'not-a-git-repo':
+                        return <p className={styles.providerExplanatoryText}>Folder is not a Git repository.</p>
+                    case 'git-repo-has-no-remote':
+                        return (
+                            <p className={styles.providerExplanatoryText}>Git repository is missing a remote origin.</p>
+                        )
+                    default:
+                        return <></>
+                }
+            } else {
+                return <></>
+            }
         default:
             return ''
     }
@@ -198,9 +215,9 @@ export const EnhancedContextSettings: React.FunctionComponent<EnhancedContextSet
             <PopupFrame
                 isOpen={isOpen}
                 onDismiss={() => setOpen(false)}
-                classNames={[popupStyles.popupTrail, styles.enhancedContextSettingsPopup]}
+                classNames={[popupStyles.popupTrail, styles.popup]}
             >
-                <div className={styles.enhancedContextInnerContainer}>
+                <div className={styles.container}>
                     <div>
                         <VSCodeCheckbox onChange={enabledChanged} checked={enabled} id="enhanced-context-checkbox" />
                     </div>